WT-7024 Move GlobalRateLimiter out of v1 directory
